# A new stack implementation that should be faster and less space-consuming